-- Change setpath submodule download without git, to not ask questions about which branch to download. Instead derive the branch name from root folder name of parent app and default to master of it cannot be determined. This change somehow did not make it earlier into repo

+++ setpaths.m
@@ -120,7 +120,9 @@ try
     appThis         = filesepStandard_startup(pwd);
     appThisPaths    = findDotMFolders(appThis, exclSearchList);
     if addremove == 0
-        deleteNamespace(appname);
+        if exist([pwd, '/Utils/namespace'],'dir')
+            deleteNamespace(appname);
+        end
         removeSearchPaths(appThis);
         return;
     end
